Fixed warning: comparison of integers of different signs when compiling with -Wsign-compare option.

+++ Classes/DDASLLogger.m
@@ -90,7 +90,7 @@ static DDASLLogger *sharedInstance;
 
         char readUIDString[16];
 #ifndef NS_BLOCK_ASSERTIONS
-        int l = snprintf(readUIDString, sizeof(readUIDString), "%d", readUID);
+        size_t l = snprintf(readUIDString, sizeof(readUIDString), "%d", readUID);
 #else
         snprintf(readUIDString, sizeof(readUIDString), "%d", readUID);
 #endif
